refactor: Move game field file to GameVO QUZ-71
