--- game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/GameLevel.kt
@@ -1,12 +1,0 @@
-package com.grepp.quizy.game.domain
-
-enum class GameLevel(
-    val description: String
-) {
-
-    EASY("쉬움"),
-    NORMAL("보통"),
-    HARD("어려움"),
-
-    ;
-}
